tcp: don't purge the error queue on sk_stream_kill_queues()
Summary:
This is the fix for socket warnings on edge.

Add it to the 5.6 branch since we're about to start fbk19.

It's not upstream yet, because we're tripple checking
it actually addresses the issue, but the fix is safe,
long story short we can't purge this queue without
holding its lock until the socket is dead-dead, and
once it's dead-dead we'll purge it again.

Test Plan: It's on 300k hosts already as hotfix2.

Reviewers: siudin

Subscribers: 

Tasks: 

Tags: 

Blame Revision: 


Differential Revision: https://phabricator.intern.facebook.com/D31660947