btrfs: skip marking metadata block groups unused
Short term, the plan for avoiding metadata under-allocation (and thus
enospc abort) is to intentionally overallocate metadata with the
force_chunk_alloc sysctl. However, if we happen to use and empty one of
the new BGs or if we remount (probably via reboot), these BGs are at
risk of being removed. To reduce churn re-allocating them constantly,
skip marking empty metadata BGS empty.

Signed-off-by: Boris Burkov <boris@bur.io>