PCI: quirks: add PCI class override for Freya
Summary:
Apply patch from Kumar Sundararajan <kumar@fb.com>:

The patch adds a Freya specific 'quirk' to pci/quirks.c

The root cause is that Freya doesn't set the class code register, it
defaults 0. This doesn't matter right after boot but Linux doesn't like
it if it finds a 0 class code when it enumerates devices.

One of the cases when this causes a problem is if we reset the PCIe
link, remove the device instance and try to add it back. We will see an
error like this:

  887.030214] freya 0000:16:00.0: can't enable device: BAR 0 [mem 0xc5d00000-0xc5dfffff] not claimed
  887.048144] freya 0000:16:00.0: Cannot enable PCI device.

Test Plan: 

Reviewers: 

Subscribers: 

Tasks: 

Tags: 

Blame Revision: 


Differential Revision: https://phabricator.intern.facebook.com/D31184314