mm: vmscan: make rotations a secondary factor in balancing anon vs file
We noticed a web RPS regression between 5.6 and 5.12. This could be
tracked down to a shift in the anon/file reclaim balance (confirmed
with swappiness) that resulted in worse reclaim efficiency and thus
more kswapd activity for the same outcome.

This is due to the anon workingset patches in 5.12. They reduced the
weight of swapins in the anon/file balance, causing more anon reclaim
than in 5.6. It's not wrong from an IO perspective, but scanning the
anon list is more expensive due to the higher ratio of mmapped pages
that may rotate during reclaim.

If rotations don't push back on the balance at all, we can end up with
very few file refaults putting all the scan pressure on hot anon pages
that are rotated en masse, don't get reclaimed, and never push back on
the file cache again. We still only reclaim file cache in that case,
but we burn a lot CPU rotating anon pages. It's "fair" from an LRU age
POV, but doesn't consider the cost it imposes on the system.

Make rotations a secondary factor in balancing the LRUs. This doesn't
attempt to make a precise comparison between IO cost and CPU cost, it
just says: if refaults/swapins are about comparable between the lists,
or rotations are overwhelmingly different, adjust for CPU work.